requirements.txt
# Core dependencies
numpy>=2.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
torch>=2.0.0,<3.0.0

# Document processing
PyPDF2>=2.12.0
python-docx>=0.8.11
markdown>=3.4.0

# RAG & LLM
langchain>=0.1.0
requests>=2.31.0

# GUI
PyQt5>=5.15.0
PyQt5-sip>=12.13.0

# Data & ML
scikit-learn>=1.3.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Export & Visualization
matplotlib>=3.7.0
reportlab>=4.0.0

# Logging & Monitoring
loguru>=0.7.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
